requirements.txt
appnope==0.1.3
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.0
decorator==4.4.2
entrypoints==0.4
executing==0.8.3
fonttools==4.37.1
gcloud==0.18.3
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-cloud-speech==2.14.1
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.3
moviepy==1.0.3
nest-asyncio==1.5.5
nltk==3.7
numpy==1.23.0
oauth2client==4.1.3
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
proglog==0.1.10
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.0
regex==2022.8.17
requests==2.28.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
selenium==4.3.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
stack-data==0.3.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
wsproto==1.1.0
